Finished removing available services as gitmodules, and setup weave.conf for app directories, available services and installed services

+++ helpers/services/commands/run.sh
@@ -5,23 +5,29 @@ set -e
 
 # This script is used to start the service
 
-echo -e "\e[33m$SERVICE_NAME: Trying to start in environment '$env_name'...\e[0m"
+echo -e "\e[33m$SERVICE_NAME: Trying to start in environment '$ENV_NAME'...\e[0m"
+
+if ! [ -f "docker-compose.yml" ]; then
+    echo -e "\e[31m$SERVICE_NAME: Cannot find docker-compose.yml file. Exiting...\e[0m"
+    exit 1
+fi
+
+if ! [ -f ".env" ]; then
+    echo -e "\e[31m$SERVICE_NAME: Cannot find .env file. Exiting...\e[0m"
+    exit 1
+fi
 
 # Prepare docker
 echo -e "\e[33m$SERVICE_NAME: Preparing docker...\e[0m"
 check_docker
 create_networks
+docker-compose down > /dev/null 2>&1
 echo -e "\e[32m$SERVICE_NAME: Docker is ready.\e[0m"
 
-# Stop existing containers
-echo -e "\e[33m$SERVICE_NAME: Stopping existing containers...\e[0m"
-docker-compose down > /dev/null 2>&1
-echo -e "\e[32m$SERVICE_NAME: Stopped existing containers.\e[0m"
-
-# Build and start containers
-if [ "$env_name" == "dev" ]; then
-    echo -e "\e[33m$SERVICE_NAME: Starting in development mode...\e[0m"
-    docker-compose -f docker-compose.yml -f docker-compose.dev.yml up --build --remove-orphans -d
+# Up docker-compose
+if [ -f "docker-compose.$ENV_NAME.yml" ]; then
+    echo -e "\e[33m$SERVICE_NAME: Starting in '$ENV_NAME'...\e[0m"
+    docker-compose -f docker-compose.yml -f docker-compose.$ENV_NAME.yml up --build --remove-orphans -d
 else
     echo -e "\e[33m$SERVICE_NAME: Starting...\e[0m"
     docker-compose up --build --remove-orphans -d
@@ -29,8 +35,6 @@ fi
 echo -e "\e[32m$SERVICE_NAME: Container started.\e[0m"
 
 # Clean up unused Docker images
-echo -e "\e[33m$SERVICE_NAME: Cleaning up unused Docker images...\e[0m"
 docker system prune -af
-echo -e "\e[32m$SERVICE_NAME: Unused Docker images cleaned up.\e[0m"
 
 echo -e "\e[32m$SERVICE_NAME: Service started successfully.\e[0m"
